ADD: id validation

+++ src/controller/book.controller.ts
@@ -7,6 +7,8 @@ import autheticate from '../middleware/authenticate.middleware'
 import authorize from '../middleware/authorize.middleware'
 import Role from '../utils/role.enum'
 import SubscriptionDto from '../dto/subscription.dto'
+import { isUUID } from 'class-validator'
+import BadRequestException from '../exception/bad-request.exception'
 
 class BookController {
   public router: Router
@@ -33,19 +35,19 @@ class BookController {
       validator(BorrowBookDto),
       this.returnBook
     ),
-    this.router.post(
-      '/:id/subscribe',
-      autheticate,
-      validator(SubscriptionDto),
-      this.addSubscription
-    ),
-    this.router.put(
-      '/:id',
-      autheticate,
-      authorize([Role.ADMIN]),
-      validator(BookDto),
-      this.updateBook
-    )
+      this.router.post(
+        '/:id/subscribe',
+        autheticate,
+        validator(SubscriptionDto),
+        this.addSubscription
+      ),
+      this.router.put(
+        '/:id',
+        autheticate,
+        authorize([Role.ADMIN]),
+        validator(BookDto),
+        this.updateBook
+      )
     this.router.delete(
       '/:id',
       autheticate,
@@ -74,7 +76,7 @@ class BookController {
         category,
         available
       )
-      
+
       res.locals.total = books.pop()
       res.locals.data = books.pop()
       res.status(200)
@@ -91,6 +93,7 @@ class BookController {
   ) => {
     try {
       const bookId = req.params.id
+      if (!isUUID(bookId)) throw new BadRequestException('Invalid book id')
 
       const book = await this.bookService.getBookById(bookId)
       res.status(200)
@@ -113,7 +116,11 @@ class BookController {
     }
   }
 
-  private addSubscription = async (req: Request, res: Response, next: NextFunction) => {
+  private addSubscription = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ) => {
     try {
       const newSubscription = await this.bookService.addSubscription(req.body)
       res.status(201)
@@ -167,6 +174,7 @@ class BookController {
   ) => {
     try {
       const bookId = req.params.id
+      if (!isUUID(bookId)) throw new BadRequestException('Invalid book id')
 
       const updatedBook = await this.bookService.updateBook(bookId, req.body)
       res.status(200)
@@ -184,6 +192,7 @@ class BookController {
   ) => {
     try {
       const bookId = req.params.id
+      if (!isUUID(bookId)) throw new BadRequestException('Invalid book id')
 
       const editedBook = await this.bookService.removeBook(bookId)
       res.status(204)
